first try at production leds
